V1.8.0
No new features but everything at the lastest version along with needed changes:
ubuntu:V24.04
Python:3.12
mammoth:V1.9.1
Sphinx:V8.2.3
sphinx_rtd_theme:3.0.2
pandoc:3.7.0.2

You need to start with a NEW project as I needed to change the tag identifier generation and the conf.py for sphinx changed as well.

The version is now displayed on the lower left corner as the current read the docs thema does not display the version on static files. Hence the workaround with a dummy version.